+++ Changes.md
@@ -1,6 +1,8 @@
 HEAD
 -----------
 
+- Worker subclasses now properly inherit sidekiq\_options set in
+  their superclass [#221]
 - Add random jitter to scheduler to spread polls across POLL\_INTERVAL
   window. [#247]